requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
altgraph==0.17.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boto3==1.26.149
botocore==1.29.149
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cssutils==2.7.1
cycler==0.11.0
dataframe-image==0.1.11
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
exceptiongroup==1.1.2
executing==1.2.0
extra-streamlit-components==0.1.56
Faker==18.10.1
fastjsonschema==2.17.1
favicon==0.7.0
filelock==3.12.2
Flask==2.2.2
flatbuffers==23.5.26
fonttools==4.39.4
fpdf==1.7.2
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
htbuilder==0.6.1
html2image==2.0.3
idna==3.4
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lightgbm==3.3.2
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
opensearch-py==2.2.0
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.7.0
pymdown-extensions==10.0.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.1
PyYAML==6.0
pyzmq==25.1.0
ray==2.6.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.11.2
selenium==4.10.0
semver==3.0.1
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
st-annotated-text==4.0.0
stack-data==0.6.2
streamlit==1.26.0
streamlit-authenticator==0.2.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.5
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webdriver-manager==3.9.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.15.0
wsproto==1.2.0
xgboost==1.5.0
yarl==1.9.2
zipp==3.15.0
